==> experiment/01_data_preparation/02_CommonNet/01_join_datasets/01_prepare_wordNet_data.ipynb <==
# Load root folder of the Git project
from pathlib import Path
import subprocess

repo_root = Path(
    subprocess.check_output(
        ["git", "rev-parse", "--show-toplevel"],
        text=True
    ).strip()
)

import pandas as pd
WORDNET_INPUT_DATASET = repo_root / "offline_datasets/processed_dataset/WordNet/common_gen_scraping.parquet"
df_wordnet = pd.read_parquet(WORDNET_INPUT_DATASET)
print(f"Loaded dataset contains {df_wordnet.shape[0]:,} rows")
df_wordnet.head(5)

### Get specific definitions > ones related to the single concepts

df_specific_definitions = df_wordnet.groupby("meanings_definitions")['concepts'].apply(lambda el: list(set(el))).reset_index()
print(f"Identified {df_specific_definitions.shape[0]:,} definitions for specific concepts")
df_specific_definitions.head(2)

### Get stats on the number of concepts per target
n_concepts_x_target = df_specific_definitions['concepts'].apply(len)
n_concepts_x_target.describe()

### Get generic definitions > ones related to the concepts' hypernyms

df_generic_definitions = df_wordnet.groupby("hypernym_definitions")['concepts'].apply(lambda el: list(set(el))).reset_index()
print(f"Identified {df_generic_definitions.shape[0]:,} definitions for broader concepts")
df_generic_definitions.head(2)

### Get stats on the number of concepts per target
n_concepts_x_target = df_generic_definitions['concepts'].apply(len)
n_concepts_x_target.describe()

### For each dataset, select a subset of data we will use only for test

# First specific definitions train/test split
import os
TRAIN_PERCENTAGE = 0.9
OUTPUT_PATH = repo_root / "offline_datasets/processed_dataset/WordNet/"
n_specific_definitions = df_specific_definitions.shape[0]
df_tmp = df_specific_definitions.copy()
df_specific_definitions_train = df_tmp.iloc[:int(n_specific_definitions*TRAIN_PERCENTAGE)].copy()
df_specific_definitions_test = df_tmp.iloc[int(n_specific_definitions*TRAIN_PERCENTAGE):].copy()
print(f"Reserved {df_specific_definitions_train.shape[0]:,} specific definitions for train and {df_specific_definitions_test.shape[0]:,} for test")
df_specific_definitions_train.to_parquet(os.path.join(OUTPUT_PATH, "specific_definitions_train.parquet"), index=False)
df_specific_definitions_test.to_parquet(os.path.join(OUTPUT_PATH, "specific_definitions_test.parquet"), index=False)

# Then, generic definitions train/test split
n_generic_definitions = df_generic_definitions.shape[0]
df_tmp = df_generic_definitions.copy()
df_generic_definitions_train = df_tmp.iloc[:int(n_generic_definitions*TRAIN_PERCENTAGE)].copy()
df_generic_definitions_test = df_tmp.iloc[int(n_generic_definitions*TRAIN_PERCENTAGE):].copy()
print(f"Reserved {df_generic_definitions_train.shape[0]:,} generic definitions for train and {df_generic_definitions_test.shape[0]:,} for test")
df_generic_definitions_train.to_parquet(os.path.join(OUTPUT_PATH, "generic_definitions_train.parquet"), index=False)
df_generic_definitions_test.to_parquet(os.path.join(OUTPUT_PATH, "generic_definitions_test.parquet"), index=False)

### Now, get all the available examples

specific_examples = df_wordnet[['concepts', 'meaning_examples']].explode('meaning_examples')
# Filter the ones in which an example is 'not available'
valid_specific_examples = specific_examples[specific_examples.meaning_examples != 'not available'].copy()
print(f"Isolated other {valid_specific_examples.shape[0]:,} examples linking a word to a specific meaning")
valid_specific_examples.head(2)

### Notice: it could be that some word shares the same example
Let's group by 'meaning_examples' and save also those examples for which it is the case!

composed_examples = valid_specific_examples.groupby('meaning_examples')['concepts'].apply(lambda el: sorted(list(set(el)))).reset_index()
# only keep samples with more than 1 concept
n_concepts_x_meaning = composed_examples.concepts.apply(len)
filtered_composed_examples = composed_examples[n_concepts_x_meaning>1].copy()
print(f"There are {filtered_composed_examples.shape[0]:,} meanings that relates to composed concepts")
filtered_composed_examples.head(2)

## Join the resulting dataset

# Convert concepts to lists
valid_specific_examples['concepts'] = valid_specific_examples['concepts'].apply(lambda el: [el])
joined_df = pd.concat([valid_specific_examples, filtered_composed_examples])
print(f"Isolated other {joined_df.shape[0]:,} examples linking set of words to a specific meaning")
joined_df.head(2)

### Now, join the three datasets composing WordNet
Also add a label, to specify that those data come from the WordNet dataset

df_specific_definitions_train.rename({'meanings_definitions':'target'}, axis=1, inplace=True)
df_generic_definitions_train.rename({'hypernym_definitions':'target'}, axis=1, inplace=True)
joined_df.rename({'meaning_examples':'target'}, axis=1, inplace=True)
wordnet_df = pd.concat([df_specific_definitions_train, df_generic_definitions_train, joined_df])
wordnet_df['dataset'] = 'WordNet'
print(f"WordNet contains {wordnet_df.shape[0]:,} elements")
wordnet_df.head(2)

### Export the new dataset

# Shuffle before saving!
SEED = 29
wordnet_df = wordnet_df.sample(frac=1, random_state=SEED)
wordnet_df.head(2)

OUTPUT_PATH = repo_root / "offline_datasets/original/CommonNet/"
OUTPUT_PATH.mkdir(parents=True, exist_ok=True)
wordnet_df.to_parquet(os.path.join(OUTPUT_PATH, "wordNet.parquet"), index=False)
