==> experiment/01_data_preparation/02_CommonNet/01_join_datasets/03_create_CommonNet.ipynb <==
# Load root folder of the Git project
from pathlib import Path
import subprocess

repo_root = Path(
    subprocess.check_output(
        ["git", "rev-parse", "--show-toplevel"],
        text=True
    ).strip()
)

### Load CommonGen

import pandas as pd
import os
INPUT_PATH = repo_root / "offline_datasets/original/CommonNet/"
common_gen_df = pd.read_parquet(os.path.join(INPUT_PATH, 'commonGen.parquet'))
print(f"Loaded dataset has {common_gen_df.shape[0]:,} elements")
common_gen_df.head(2)

### Load WordNet

wordNet_df = pd.read_parquet(os.path.join(INPUT_PATH, 'wordNet.parquet'))
print(f"Loaded dataset has {wordNet_df.shape[0]:,} elements")
wordNet_df.head(2)

### Join datasets

full_df = pd.concat([common_gen_df, wordNet_df])
print(f'Full dataset has {full_df.shape[0]:,} rows')
full_df.head(2)

### Shuffle before exporting

SEED = 29
shuffled_df = full_df.sample(frac=1, random_state=SEED)
shuffled_df.head(2)

### Out of curiosity: get some stats on the number of concepts 

n_concepts = shuffled_df.concepts.apply(lambda list_concepts: len(list_concepts))
n_concepts.describe()

### Filtering the dataset
Remove rows with a number of concepts > 5 as the network will not going to learn those examples 


filtered_df = shuffled_df[n_concepts<=5].copy()
print(f'Dataset filtered to {filtered_df.shape[0]:,} rows (it had {shuffled_df.shape[0]:,})')
filtered_df.head(2)

### Check again the stats

n_concepts = filtered_df.concepts.apply(lambda list_concepts: len(list_concepts))
n_concepts.describe()

### Export the filtered dataset to a parquet file

import os
OUTPUT_PATH = repo_root / "offline_datasets/original/CommonNet/"
OUTPUT_PATH.mkdir(parents=True, exist_ok=True)
filtered_df.to_parquet(os.path.join(OUTPUT_PATH, 'commonNet.parquet'), index=False)
